[refactor] 카카오 로그인 의존성 주입 관련 수정
