+++ app/src/main/java/org/go/sopt/winey/di/ServiceModule.kt
@@ -1,15 +1,11 @@
 package org.go.sopt.winey.di
 
-import android.content.Context
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.user.UserApiClient
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.go.sopt.winey.data.service.AuthService
 import org.go.sopt.winey.data.service.FeedService
-import org.go.sopt.winey.data.service.KakaoLoginService
 import org.go.sopt.winey.data.service.NotificationService
 import org.go.sopt.winey.data.service.RecommendService
 import retrofit2.Retrofit
@@ -40,47 +36,4 @@ object ServiceModule {
     @Singleton
     fun provideNotificationService(retrofit: Retrofit): NotificationService =
         retrofit.create(NotificationService::class.java)
-
-    @Provides
-    fun provideKakaoLoginService(
-        client: UserApiClient
-    ): KakaoLoginService {
-        return object : KakaoLoginService {
-            override fun loginKakao(
-                kakaoLoginCallBack: (OAuthToken?, Throwable?) -> Unit,
-                context: Context
-            ) {
-                val kakaoLoginState =
-                    if (client.isKakaoTalkLoginAvailable(context)) {
-                        KAKAO_TALK_LOGIN
-                    } else {
-                        KAKAO_ACCOUNT_LOGIN
-                    }
-
-                when (kakaoLoginState) {
-                    KAKAO_TALK_LOGIN -> {
-                        client.loginWithKakaoTalk(
-                            context,
-                            callback = kakaoLoginCallBack
-                        )
-                    }
-
-                    KAKAO_ACCOUNT_LOGIN -> {
-                        client.loginWithKakaoAccount(
-                            context,
-                            callback = kakaoLoginCallBack
-                        )
-                    }
-                }
-            }
-
-            override fun logoutKakao(kakaoLogoutCallBack: (Throwable?) -> Unit) {
-                client.logout(kakaoLogoutCallBack)
-            }
-
-            override fun deleteKakaoAccount(kakaoLogoutCallBack: (Throwable?) -> Unit) {
-                client.unlink(kakaoLogoutCallBack)
-            }
-        }
-    }
 }
